clean up firefoxhomeviewcontroller and library creation and usage
