--- Client/Frontend/Home/HomeViewController.swift
@@ -1,13 +1,0 @@
-//
-//  HomeViewController.swift
-//  Client
-//
-//  Created by Krzysztof Modras on 26.08.19.
-//  Copyright © 2019 Cliqz. All rights reserved.
-//
-
-class HomeViewController: FirefoxHomeViewController, HomeViewControllerProtocol {
-    func scrollToTop() {
-        super.scrollToTop()
-    }
-}
